+++ sdks/web-ui-sdk/CHANGELOG.md
@@ -1,5 +1,13 @@
 # web-ui-sdk
 
+## 1.5.60
+
+### Patch Changes
+
+- core
+- Updated dependencies
+  - @ballerine/common@0.9.59
+
 ## 1.5.59
 
 ### Patch Changes